FR-1453 Reorganize functions into state_machine.go and helper.go
